[Tilma][Bristol] Add pre-filter option when calling layer
As the layer contains a lot of data we are not using we
need to set up a call for what we need instead
of filtering down the results we get as that is going
to be resource intensive.

This is paired with servers branch bristol-alloy-staging-assets
which uses the new config.

(That branch also removes some unused alloy configuration)

https://3.basecamp.com/4020879/buckets/39557940/todos/8032010747