[Tilma] Download srwr data and process for streetmanger
Download the daily srwr data and process the provided
csv file to replicate the fields for the Streetmanager
database and insert the data into that.

Currently a work in progress - start and end dates need
to be checked against phases and perhaps the entries
dropped if there is no start and end field.

Some fields currently have hardcoded data that should
be parsed from the csv data.

https://github.com/mysociety/societyworks/issues/4733